third_party/blink/web_tests/fast/js/webcore-string-comparison-expected.txt

Test for bug 33057: typeof(xhr.responseText) != "string" on Windows.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof(r.responseText) == "string" is true
PASS typeof(r.responseText) === "string" is true
PASS typeof(r.responseText) != "string" is false
PASS typeof(r.responseText) !== "string" is false
PASS "" + typeof(r.responseText) is "string"
PASS typeof(r.responseText) + "" is "string"
PASS String(typeof(r.responseText)) is "string"
PASS String(typeof(r.responseText)) is typeof(r.responseText)
PASS successfullyParsed is true

TEST COMPLETE
